Add MIT License
Uploaded the MIT License file to the project repository. The MIT License grants users permission to freely use, modify, and distribute the software, subject to certain conditions. By including this license, we provide clarity on how others can use our code and contribute to the project.

The license text includes copyright information for the year 2024 and the name "Relangi Tarun Kumar." This ensures that proper attribution is maintained for the project.

The license file was added to the root directory of the project repository using the GitHub web interface. It can now be accessed and viewed by anyone visiting the repository.

This addition aligns the project with best practices for open-source software development and encourages collaboration within the community.